Move Meeting class into meeting.meeting package
