+++ src/main/java/seedu/address/logic/commands/meeting/AddCommand.java
@@ -4,8 +4,8 @@ import static java.util.Objects.requireNonNull;
 
 import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.exceptions.CommandException;
-import seedu.address.model.meeting.Meeting;
 import seedu.address.model.meeting.ModelMeeting;
+import seedu.address.model.meeting.meeting.Meeting;
 
 public class AddCommand extends Command {
     public static final String COMMAND_WORD = "add";
