+++ src/main/java/seedu/address/logic/parser/meeting/AddCommandParser.java
@@ -16,7 +16,7 @@ import seedu.address.logic.parser.ArgumentTokenizer;
 import seedu.address.logic.parser.Parser;
 import seedu.address.logic.parser.Prefix;
 import seedu.address.logic.parser.exceptions.ParseException;
-import seedu.address.model.meeting.Meeting;
+import seedu.address.model.meeting.meeting.Meeting;
 
 /**
  * Parses input arguments and creates a new AddCommand object
